added deleting functions for all elements , reinitialized, added simulator , fixed issues

+++ DalXml/DalXml/DalXml.cs
@@ -60,10 +60,10 @@ namespace Dal
         }
 
         [MethodImpl(MethodImplOptions.Synchronized)]
-        public double[] DronePwrUsg()
+        public int[] DronePwrUsg()
         {
             XElement tempPwr = XMLTools.LoadListFromXMLElement(@"config.xml");
-            double[] pwrUsg = { int.Parse(tempPwr.Element("pwrUsgEmpty").Value), int.Parse(tempPwr.Element("pwrUsgLight").Value), int.Parse(tempPwr.Element("pwrUsgMedium").Value), int.Parse(tempPwr.Element("pwrUsgHeavy").Value), int.Parse(tempPwr.Element("chargePH").Value) };
+            int[] pwrUsg = { int.Parse(tempPwr.Element("pwrUsgEmpty").Value), int.Parse(tempPwr.Element("pwrUsgLight").Value), int.Parse(tempPwr.Element("pwrUsgMedium").Value), int.Parse(tempPwr.Element("pwrUsgHeavy").Value), int.Parse(tempPwr.Element("chargePH").Value) };
             return pwrUsg;
         }
